refactor: stop importing scanpy's private plotting internals
cellrank pulled two private symbols out of scanpy — `palettes.vega_20_scanpy` and
`_utils.add_colors_for_categorical_sample_annotation` — both of which scanpy 1.13
relocated to `scanpy.plotting.legacy`, breaking collection of the whole suite on the
pre-release job.

Neither needs to come from scanpy: the palette is a fixed reordering of matplotlib's
`tab20`, now defined (and pinned by a test) in `_utils/_colors.py`, and the colour
assignment is what cellrank's own `_get_categorical_colors` already does for the
plotting functions. That helper moves from `pl/_utils` to `_utils/_colors` so
`models/` can use it without importing from `pl/` and now also regenerates palettes
that are too short or not colour-like, which is the part scanpy's helper contributed.
