+++ src/cellrank/pl/_log_odds.py
@@ -14,10 +14,11 @@ from matplotlib.cm import ScalarMappable
 from matplotlib.colors import Normalize, to_hex
 
 from cellrank._utils import Lineage
+from cellrank._utils._colors import _get_categorical_colors
 from cellrank._utils._docs import d
 from cellrank._utils._utils import _unique_order_preserving, save_fig
 from cellrank.kernels._utils import _ensure_numeric_ordered
-from cellrank.pl._utils import _get_categorical_colors, _position_legend
+from cellrank.pl._utils import _position_legend
 
 logger = logging.getLogger(__name__)
 __all__ = ["log_odds"]
