+++ tests/test_colors.py
@@ -30,6 +30,32 @@ class TestColors:
         assert all(isinstance(c, str) for c in colors), colors
         assert all(is_color_like(c) for c in colors), colors
 
+    def test_create_categorical_colors_first_20_are_pinned(self):
+        # cellrank owns this palette (it used to come from `scanpy.plotting.palettes`);
+        # the committed ground-truth figures depend on these exact values
+        assert _create_categorical_colors(20) == [
+            "#1f77b4",
+            "#ff7f0e",
+            "#279e68",
+            "#d62728",
+            "#aa40fc",
+            "#8c564b",
+            "#e377c2",
+            "#b5bd61",
+            "#17becf",
+            "#aec7e8",
+            "#ffbb78",
+            "#98df8a",
+            "#ff9896",
+            "#c5b0d5",
+            "#c49c94",
+            "#f7b6d2",
+            "#dbdb8d",
+            "#9edae5",
+            "#ad494a",
+            "#8c6d31",
+        ]
+
 
 class TestMappingColors:
     def test_mapping_colors_not_categorical(self):
